Tornado as a monster spell for the sole use of Lom Lobon.
This takes the place of one instance of Blink Range and instead of the
emergency Ice Storm.

Currently this isn't nerfed by monster translocations; I question if we
really want to have Blink Range on Lom anyway (it makes him extremely
easy to avoid).

If there's another scenario in which this is likely to work, well, it's
here now.