Asterion, a late-game unique minotaur of Makhleb
He is themed as a bloodthirsty ruler cursed to become a minotaur who now wanders
the dungeon in the service of Makhleb and hoping to establish a new kingdom in
its depths.

The name is taken from ancient greek mythology, which was either the name of the
minotaur itself or of a king related to the minotaur:

http://en.wikipedia.org/wiki/Asterion

Thanks to SGrunt (the S is silent) for the idea of the name.

Stat-wise, Asterion is a minotaur, but with 20 HD, 175 HP, a 1d40 primary attack
(instead of 1d35), and the only resistances I've added are sinv and high MR.
Experience and placement-wise he's the same as sojobo. He always gets a good
great mace and good armour, either chain mail or plate.

Spell-wise, he has the two high-level Makhleb invocations: major destruct and
greater servant.  The major destruct code is basically identical to the player
version: it chooses randomly from 7 beam types, including orb of electricity,
with a range of 7. The greater servant summons a high-level demon from the same
list as the player ability: executioner, green death, blizzard demon, balrug, or
cacodemon, all equally likely.  This monster version of greater servant uses the
summons cap system with a cap of two; if Asterion makes a third servant, the one
with the shortest remaining duration will time out the next turn.

The spell messages are themed to invoke Makhleb, but I haven't given him any
speech text yet.