Remove Demonic Horde
This was neither a popular nor powerful nor especially interesting
spell, which was basically 'Call Imp, except a lot more of them'.
Making the demons more individually powerful only further steps on
the toes of the many other demon summoning spells that already
exist, so I think it is fine to simply trim it.