Consider a band produced by shadow creatures to be one creature for capping purposes
This means that each new cast of shadow creatures can replace only
one previous cast of shadow creatures, instead of pulling a band of
green rats and overwriting a hydra and 2 death yaks with them (only
for the band members to dissolve on their own shortly thereafter
anyway). I think this will make it feels a little more 'consistent'
(or at least as much as one can expect from such a random spell)

The whole band is considered both as a single creature when counting
whether the summoner is at their cap, and will also always expire
together when any of them are removed due to capping.