Reduce Summon Mana Viper's level to 5
It was generally harder to cast than Summon Menagerie, yet less
powerful in many contexts. Mana vipers do provide stronger disabling
support, but this is less needed if you can just kill the target
instead.

The spell might benefit from not being found only in a book that also
has Menagerie, incidentally - maybe it should be put in one of the
hex books?