Give monsters a chance to drop meat perma-food on death
A 1/4 chance to drop beef jerky or meat ration on death, strongly
weighted towards jerky, if the monster has normal intelligence or
higher, leaves a corpse, and the player isn't worshiping gozag.

This gives an average drop of 550 nutrition per monster able to
generate the food. It's a bit harder to estimate the average nutrition
previously received from chunks given the number of variables, but
assuming monsters drop 2 chunks on average, and assuming a nutrition
weighted toward contaminated chunks at 625 per chunk (3:1 contaminated
vs. not-contaminated), monsters drop 625 nutrition on average through
their corpse with a 50% corpse drop rate. This is probably an
underestimate of that average, but we certainly aren't dropping more
nutrition this way compared to what we did through chunks.