Rework the player version of simulacrum
It now behaves like animate skeleton, creating simulacra from the top
corpse the player is standing on and leaving a skeleton (but not a
hide) when appropriate. The number created is based on how many chunks
would be produced, so the result is the same as butchering a corpse
and wielding all the resulting chunks under the old version of the
spell.