Make auto_eat_chunks and easy_eat_chunks automatically eat rations
You will automatically eat a piece of food such that it won't waste any
nutrition.

(Keeping the options the same long term is probably a bad idea, but for
now this keeps compatibility with Trunk rc files. Separating the options
can happen if/when chunkless is ready to be moved to trunk. -reaverb)