Remove the chance of monsters dropping beef jerky
This serves two purposes:

A) Less annoying because beef jerky is the lesser nutrition

B) Testing a severe nerf of the food weight.

A big change to the system used to distribute food is coming within the
next week or so, but these changes might as well gather data now.