Pakellas: Quick Charge.
Spend your entire MP pool to charge, on average, (10 * evocations) *
(MP / max MP) percent of a wand or rod's maximum charges to the
wand/rod. Does not upgrade rods, but does charge them.

The fact that most means of MP recovery are blocked is intended to
specifically balance this ability.