Remove the hunger/regen cap on movement
Previously, various over-time effects behaved oddly with slow
movement. When walking, any action that took more than 10 aut
(one normal turn) was counted as if it took 10 aut instead for
regen/hunger purposes. This meant that, for example, moving in
statue form cut your hunger & regen to 2/3rds normal; Chei
worshippers had both halved...

This removes that effect and adds a few small compensations;
Naga get Slow Metabolism 1 and Chei reduces hunger significantly
more. (Now cuts it to 50%; previously 75%).

There's no save compat code in this commit for existing naga (to
give them the mut); if that's felt to be important, we can add a
tag, but I'm trying to be conservative with those.