Trigger travis-ci rebuild
Debugging whether a test failure is caused by this branch being pushed
to master. (See: https://travis-ci.org/crawl/crawl/jobs/60096027).
There's probably a better way to do this but it's easy just to push a
commit.

Previous travis-ci build was contaiminated by a force push (causing
most of the build to fail)