Replace the RNG with ChaCha
What's ChaCha? It's the RNG you should be using. It's a stream cipher designed to resist
timing attacks. This really isn't relevant to crawl, but let's not set a bad example.

Why bother? The alternating step generator KISS RNG is probably unfit for any purpose.

What the heck is going on in random.cc for real number generation?
Magic. The distance between values on the IEEE floating point number line is
uniform between [1,2). To exploit this, generate a number from [1,2), then
subtract 1. By the power of magic, you get a pretty darn good random number
on the interval [0,1). If anyone finds this objectionable, feel free to replace it.

Should this be shoved into master?
No. I need to run it through the paces. There's probably some catastrophic bug
lurking in here.