Use X-Forwarded-Proto header for websocket protocol
If the webtiles server is running behind a proxy server
which terminates SSL connections, and http is used between
the proxy and webtiles, the X-Forwarded-Proto header can be
set by the proxy to indicate to the webtiles server that the
client is using a secure connection. This allows webtiles to
generate the correct protocol wss:// even it sees http.