Add Matthew Ludivico to CREDITS.txt
Move his name properly into the credits instead of having it in various
comments for vault files. His name is also credited in the names of the
vaults themselves.

His name was listed in comments at the top of several vaults in addition
to his handle being in the names of said vaults.