Adjust ossuary_the_hunt_dpeg
This ossuary has not worked for some time: zombies can't open doors. The
revelation of traps has made it even worse, as the "nasty" trap mode can
just be avoided by stepping around them.

The doors are removed and trap selection is revised, removing most of
the mechanicals. Net traps are now player-targeting, so are the lone
mechanical trap that is kept.