Display real MP for Deep Dwarves when necessary (Ge0ff)
Deep Dwarves' Heal Wounds ability is the only one that requires real MP.
Even if the maximum MP is 0 because of an antimagic weapon or -MP items,
the player can still use the ability. Similarly, if the player has no
real magic points, but has +MP items, the ability cannot be used.

Currently it is not possible to check the real MP without taking
actions. Taking off items with +MP and -MP or switching weapons takes
time, and unequpping some items, e.g. Maxwell's etheric cage, has side
effects.

This patch adds real MP to the %-screen/morgue and to the HUD (only for
Deep Dwarves and only when real MP != max MP).