A Klonw Kar trap for Zot
Generalize the basket of spiders trap to be a monster-drop trap.
Then add the trap to Zot but have it drop Killer Klowns. Also change the
snake version to drop a mix of mambas and mana vipers.

The basket traps work ok as tactical terrain even with traps revealed,
since a monster triggering the trap will still release a player-targeting
effect (more monsters).