Reduce silence radius from Klown pies
Monster silence is replaced with mute, as monster silence is full radius
(5) for all but the last 3 turns of its duration. Player silence is
reduced to 4+random2(7) turns, with a cap of 10. This means the player
will have at most four turns with a silence radius of 1, followed by 6
turns of radius 0.

The silence effect is strong, and primarily aimed at disabling the
player. The monster silence effect meant that instead of being able to
use a consumable or be stuck a player could be left in an awkward
situation of wanting to kill an ally to break the silence. On the flip
side, silence is strong against ancient liches and the draconian
spellcasters in the zot monster set. The anti-synergy between klowns and
these monsters hass the potential for if not abuse, at least strange
incentives.

Reducing the silence duration is also a nerf to the new klown pie
effect; maintaining the old duration without having a large silence
radius would require a new status effect.