Convert `enum branch_flag_type` to `enum class brflag`
The name of the enum class changes to be significantly shorter, so that
references to members of the enum class are of a similar length to
before. Additionally, add a new bitfield type to use for the member
variable Branch::branch_flags - this type is called branch_flags_t.

Convert some uses of bitwise and to use testbits, where the raw bitwise
and will now fail to compile (in return statements and variable
definitions).