Editing/reordering pass for the species appendix of the manual
I'm not sure what previously determined the order of species in this
section, but it was kind of idiosyncratic. I have reordered the species
to match the order on the newgame menu, and explicitly included the
categories used in selecting a new character. I also took an editing
pass over the text and made a few tweaks that hopefully will address
common questions about species selection, including the misconception
that order on that menu is about difficulty of winning.

(An alternative would have been to simply alphabetize this section, but
probably the most likely case where players will look at it, at least
after the fix in 146c4415506c6f, is when hitting ? from this menu.)