Indicate spell damage with exclamation marks and "...but does no damage"
Use varying numbers of exclamation marks to indicate the amount of
damage done by direct-damage spells that pass through zapping(), in the
same way as is done for melee and ranged attacks, and print
   "<beam> hits <actor> but does no damage."
to the message log when a beam does no damage, in the same way as is
done for melee attacks.

This behaviour was previously inconsistent: some beam types printed
"%s appears unharmed" for monsters when they were hit for no damage,
others simply printed nothing, and this message was never printed
when the player was hit by a beam for no damage. These inconsistent
"%s appears unharmed" messages are removed in this commit.

This largely eliminates a long-running inconsistency in the output of
spells vs melee attacks.

There are various other instances of spell damage messages scattered
throughout the code that should also be converted in the future.
These include torment effects, damage-over-time effects, and clouds.

Also remove the "Helpless, you fail to dodge..." message for when you're
paralysed, because there is no clean way to include this in the new
system, it doesn't say anything especially meaningful (the player knows
they're paralysed), and it also awkwardly triggered when in tree form.

Finally, remove an unnecessary variable and replace it with explicit
return true; / return false; statements where it was used.