Add trap placement to Tomb:1 hallway end vaults
The Tomb:1 hallway end vaults contain the arrival hatch into the Tomb:1
hall from Tomb:2 at one end and the exit hatch into the second section
of Tomb:2 on the other end. In these vaults add small numbers of traps
as appropriate for their smaller size, about 6 on average. Also reglyph
the vaults to use the usual '.' for floor, as there's no need to use a
special glyph.