Rework trap placement on Tomb:3
In the large central Tomb:3 vaults, place about 12 traps on-average.
This is enough to add some hindrance for the player as they move and fight
through the central area without being overwhelming.

Mechanical traps were removed from the treasure rooms on Tomb:3, but
numbers of traps as a whole weren't changed, so alarm/dispersal/zot
traps became 2.5 times as common. This commit reduces numbers of traps
in the treasure rooms overall, still leaving a higher trap density
compared to the central rune areas. This trap placement can now also
occur in the initial hallway where the player arrives on Tomb:3.

Also use normal '.' glyphs for the treasure room and refactor DES of the
main map.