Fix a debug warning message
This debug message wasn't properly updated in a6713cf8 and causes a
compilation warning due to implicit conversion of an enum class to
unsigned int. Use a static cast and have the message attempt to also
display the spell school's long name. It might be better to just use an
ASSERT here, but I haven't looked at what all the callers are doing with
this function.