Don't travel through excluded transporters (EN0N)
Currently exclusions made on transporters do nothing to limit travel.
Travel handles transporters like stairs in terms of allowing the player
to move onto the excluded transporter if the entire exclusion is a
single square. This is good, since just as we do for single-square
exclusions on stairs, we want to prevent travel from taking the
transporter but allow standing on it. However travel didn't actually
prevent taking the transporter, and with this commit it now does.