Rework a loot item in a vault
Replace the "acquire any" items in evilmike_wizard_prison with nice
randart artefact jewelry, randart aux armour, or acquirement scrolls.
The scroll is a better way to give player loot that's based on their
character, since the choice aspect makes it more likely to be useful.

Also fix the wrong usage of a substitution operator and simplify terms
in some NSUBST.