Use a stable order for loading vault files
It turns out that all the deterministic RNG in the world doesn't help if
you are choosing among things whose initial order is at the mercy of the
file system.

This commit isn't necessarily a full and complete solution to this
issue, more of a minimal first step; it would be nice to have an even
more stable way of getting vault indices.