Give Tengu permanent flight at XL 5
This was suggested on Mantis as 10707 and a patch was included,
unfortunately the patch had aged poorly and couldn't cleanly apply, so
this is a re-implementation.

Tengu racial flight starts at XL 5, as an activated ability. The
activation gives flight (with its Tengu bonuses) for 25 + random2(xl*4)
+ random2(xl*4) turns. In practice, this meant that a low XL Tengu
effectively had permanent flight, provided they remembered to activate
it. The MP cost takes at most 38 turns to recover (regen to full with 3
MP). The duration of Tengu flight averaging 45 turns when first
acquired, rising to 77 at XL 13. The Tengu player with more than 3
total MP that was willing to do a lot of tedious waiting and re-casting
could enter any encounter flying and with full MP.

Removing the interface screw shouldn't affect optimal balance, but it
might turn out to be a practical buff. If that haappens to be the case a
nerf that is not an interface screw can be applied.