Make Jiyva slime neutralising like Fedhas' (11792)
Changing god conversions to run exactly once, and on first sight
resulted in unkillable slimes for a Jiyva player. This moves slime
conversion to run in the same place as Fedhas' plant neutralising,
making it global and independent of player LOS.

This has the effect of waking up slimes on a level so they start eating
items sooner; however awake slimes that hadn't been converted would
start eating already, so the net effect isn't that large.